+++ modules/aws/core/main.tf
@@ -6,7 +6,7 @@
   */
 
 terraform {
-  required_version = "1.1.7"
+  required_version = ">= 1.1.7"
   required_providers {
     aws = {
       source  = "hashicorp/aws"