+++ modules/azure/github-runner/main.tf
@@ -11,7 +11,7 @@
   */
 
 terraform {
-  required_version = "1.1.7"
+  required_version = ">= 1.1.7"
 
   required_providers {
     azurerm = {