+++ modules/kubernetes/azad-kube-proxy/main.tf
@@ -62,7 +62,7 @@
 */
 
 terraform {
-  required_version = "1.1.7"
+  required_version = ">= 1.1.7"
 
   required_providers {
     kubernetes = {