+++ modules/kubernetes/grafana-agent/main.tf
@@ -54,7 +54,7 @@
 */
 
 terraform {
-  required_version = "1.1.7"
+  required_version = ">= 1.1.7"
 
   required_providers {
     kubernetes = {