+++ modules/azure/aks/main.tf
@@ -13,7 +13,7 @@
   */
 
 terraform {
-  required_version = "0.15.3"
+  required_version = ">= 1.1.7"
 
   required_providers {
     azurerm = {