+++ modules/kubernetes/opa-gatekeeper/main.tf
@@ -85,7 +85,7 @@
   */
 
 terraform {
-  required_version = "0.15.3"
+  required_version = ">= 1.1.7"
 
   required_providers {
     kubernetes = {