+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [#608](https://github.com/XenitAB/terraform-modules/pull/608) Include node-local-dns IP in default-deny networkpolicy CIDR block.
 - [#611](https://github.com/XenitAB/terraform-modules/pull/611) Add option to turn off ingress-nginx metrics in grafana.
 
+### Changed
+
+- [#607](https://github.com/XenitAB/terraform-modules/pull/607) [Breaking] Upgrade terraform to 1.1.7.
+
 ## 2022.03.3
 
 ### Changed